+++ CSharp10.Tests/UnitTests/ConstantInterpolatedStringTests.cs
@@ -1,4 +1,4 @@
-﻿namespace PolySharpFeatureTests.CSharp10;
+﻿namespace CSharp10.Tests.UnitTests;
 
 [TestClass]
 public class ConstantInterpolatedStringTests
