+++ CSharp10.Tests/UnitTests/ExtendedPropertyPatternTests.cs
@@ -1,4 +1,4 @@
-﻿namespace PolySharpFeatureTests.CSharp10;
+﻿namespace CSharp10.Tests.UnitTests;
 
 [TestClass]
 public class ExtendedPropertyPatternTests
@@ -21,13 +21,17 @@ public class ExtendedPropertyPatternTests
 
     private class Foo
     {
+#pragma warning disable IDE0090 // Use 'new(...)'
         public Bar Bar { get => new Bar(); }
+#pragma warning restore IDE0090 // Use 'new(...)'
         public string Value = "Foo";
     }
 
     private class Bar
     {
+#pragma warning disable IDE0090 // Use 'new(...)'
         public Baz Baz { get => new Baz(); }
+#pragma warning restore IDE0090 // Use 'new(...)'
         public string Value = "Bar";
     }
